[minor] Inline {BufferUtil,Validation}.fallback.js into parent modules
