+++ lib/Validation.js
@@ -13,5 +13,5 @@ try {
     ? isValidUTF8.Validation.isValidUTF8  // utf-8-validate@<3.0.0
     : isValidUTF8;
 } catch (e) {
-  module.exports = require('./Validation.fallback');
+  module.exports = () => true;
 }
